+++ src/main/java/com/aquavitae/domain/repository/EstadoPlantaRepository.java
@@ -3,5 +3,5 @@ package com.aquavitae.domain.repository;
 import java.time.LocalDateTime;
 
 public interface EstadoPlantaRepository {
-    void guardarEstado(int plantaId, float indiceHidrico, LocalDateTime fecha);
+    void guardarEstado(int plantaId, Double indiceHidrico, LocalDateTime fecha);
 }
